Retrieve palmitoylation data from the swissPalm database

Retrieve palmitoylation data from the swissPalm database

Usage

swissPalm(query_id, dataset = "all", species = "")

Arguments

query_id

Character vector with protein identifiers. Should be an ID type supported by SwissPalm (UniProt AC, UniProt secondary AC, UniProt ID, UniProt gene name, Ensembl protein, Ensembl gene, Refseq protein ID, IPI ID, UniGene ID, PomBase ID, MGI ID, RGD ID, TAIR protein ID, EuPathDb ID; from the SwissPalm website).

dataset

Which dataset to use in swissPalm. Will be set to "all" if not a value from swisspalm::datasets. Default = "all" (all datasets).

species

Which species to use in swissPalm. Will be set to "" if not a value from swisspalm::species. Default = "" (all species).

Value

Data frame with palmitoylation data for proteins in protein_id, and notes for why any elements in protein_id were not found in SwissPalm.

Examples

# \donttest{
swisspalmR::swissPalm(query_id = c("P05067", "O00161", "P04899", "P98019"))
#>   Query_identifier UniProt_AC  UniProt_ID UniProt_status           Organism
#> 1           P05067     P05067    A4_HUMAN       Reviewed       Homo sapiens
#> 2           P04899     P04899 GNAI2_HUMAN       Reviewed       Homo sapiens
#> 3           O00161     O00161 SNP23_HUMAN       Reviewed       Homo sapiens
#> 4           P98019     P98019  COX2_ANAPL       Reviewed Anas platyrhynchos
#>                   Gene_names
#> 1               AD1, A4, APP
#> 2              GNAI2B, GNAI2
#> 3                     SNAP23
#> 4 MTCO2, COXII, COII, MT-CO2
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                Description
#> 1 Amyloid-beta precursor protein, APP (ABPP) (APPI) (Alzheimer disease amyloid A4 protein homolog) (Alzheimer disease amyloid protein) (Amyloid precursor protein) (Amyloid-beta (A4) precursor protein) (Amyloid-beta A4 protein) (Cerebral vascular amyloid peptide, CVAP) (PreA4) (Protease nexin-II, PN-II) [Cleaved into: N-APP; Soluble APP-alpha, S-APP-alpha; Soluble APP-beta, S-APP-beta; C99 (Beta-secretase C-terminal fragment, Beta-CTF); Amyloid-beta protein 42, Abeta42 (Beta-APP42); Amyloid-beta protein 40, Abeta40 (Beta-APP40); C83 (Alpha-secretase C-terminal fragment, Alpha-CTF); P3(42); P3(40); C80; Gamma-secretase C-terminal fragment 59 (Amyloid intracellular domain 59, AICD-59, AID(59)) (Gamma-CTF(59)); Gamma-secretase C-terminal fragment 57 (Amyloid intracellular domain 57, AICD-57, AID(57)) (Gamma-CTF(57)); Gamma-secretase C-terminal fragment 50 (Amyloid intracellular domain 50, AICD-50, AID(50)) (Gamma-CTF(50)); C31 ]
#> 2                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                   Guanine nucleotide-binding protein G(i) subunit alpha-2 (Adenylate cyclase-inhibiting G alpha protein)
#> 3                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    Synaptosomal-associated protein 23, SNAP-23 (Vesicle-membrane fusion protein SNAP-23)
#> 4                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         Cytochrome c oxidase subunit 2, EC 7.1.1.9 (Cytochrome c oxidase polypeptide II)
#>   Number_of_palmitoyl_proteomics_articles
#> 1                                 2 of 22
#> 2                                17 of 22
#> 3                                20 of 22
#> 4                                  0 of 0
#>   Number_of_palmitoyl_proteomics_studies_where_the_protein_appears_in_a_high_confidence_hit_list
#> 1                                                                                              3
#> 2                                                                                             11
#> 3                                                                                             11
#> 4                                                                                              0
#>   Number_of_technique_categories_used_in_palmitoyl_proteomics_studies
#> 1                                                                   2
#> 2                                                                   4
#> 3                                                                   4
#> 4                                                                   0
#>                                                                                                            Technique_categories_used_in_palmitoyl_proteomics_studies
#> 1                                                               Chemical modification and capture of palmitoylated cysteines, Identification of palmitoylation sites
#> 2 Chemical modification and capture of palmitoylated cysteines, Identification of PATs/APTs, Identification of palmitoylation sites, Metabolic labeling of palmitate
#> 3 Chemical modification and capture of palmitoylated cysteines, Identification of PATs/APTs, Identification of palmitoylation sites, Metabolic labeling of palmitate
#> 4                                                                                                                                                                 NA
#>   Number_of_targeted_studies                       Targeted_studies__PMIDs
#> 1                          4                                      23825420
#> 2                          7 31142470, 19001095, 8475115, 9892647, 7651161
#> 3                          4                            31142470, 10329400
#> 4                          0                                            NA
#>                             PATs   APTs Number_of_sites   Sites_in_main_isoform
#> 1 Q91WU6, Q9D270, Q9NXF8, Q8IVQ6     NA               2                186, 187
#> 2 Q8R173, Q91WU6, P59267, Q9D270 O75608               3            66, 352, 112
#> 3                             NA     NA               6 79, 80, 83, 85, 87, 112
#> 4                             NA     NA               0                      NA
#>   Number_of_isoforms Max_number_of_cysteines
#> 1                 11                      18
#> 2                  6                      11
#> 3                  2                       6
#> 4                  1                       3
#>   Max_number_of_cysteines_in_TM_or_cytosolic_domain
#> 1                                                 0
#> 2                                                10
#> 3                                                 6
#> 4                                                NA
#>   Predicted_to_be_S_palmitoylated
#> 1                            TRUE
#> 2                            TRUE
#> 3                            TRUE
#> 4                           FALSE
#>   Predicted_to_be_S_palmitoylated_in_cytosolic_domains
#> 1                                                 TRUE
#> 2                                                 TRUE
#> 3                                                 TRUE
#> 4                                                FALSE
#>   Protein_has_hits_in_SwissPalm
#> 1                          TRUE
#> 2                          TRUE
#> 3                          TRUE
#> 4                         FALSE
#>   Orthologs_of_this_protein_have_hits_in_SwissPalm Found_in_SwissPalm
#> 1                                             TRUE              found
#> 2                                             TRUE              found
#> 3                                             TRUE              found
#> 4                                            FALSE              found
# }

# Use 'species' parameter to limit your results to those for a species of 
# interest
# \donttest{
swisspalmR::swissPalm(
  query_id = c("P05067", "O00161", "P04899", "P98019"),
  species = swisspalmR::species["Mallard duck"]
  )
#>   Query_identifier UniProt_AC UniProt_ID UniProt_status           Organism
#> 1           P98019     P98019 COX2_ANAPL       Reviewed Anas platyrhynchos
#> 2           O00161       <NA>       <NA>           <NA>               <NA>
#> 3           P04899       <NA>       <NA>           <NA>               <NA>
#> 4           P05067       <NA>       <NA>           <NA>               <NA>
#>                   Gene_names
#> 1 MTCO2, COXII, COII, MT-CO2
#> 2                         NA
#> 3                         NA
#> 4                         NA
#>                                                                        Description
#> 1 Cytochrome c oxidase subunit 2, EC 7.1.1.9 (Cytochrome c oxidase polypeptide II)
#> 2                                                                             <NA>
#> 3                                                                             <NA>
#> 4                                                                             <NA>
#>   Number_of_palmitoyl_proteomics_articles
#> 1                                  0 of 0
#> 2                                    <NA>
#> 3                                    <NA>
#> 4                                    <NA>
#>   Number_of_palmitoyl_proteomics_studies_where_the_protein_appears_in_a_high_confidence_hit_list
#> 1                                                                                              0
#> 2                                                                                             NA
#> 3                                                                                             NA
#> 4                                                                                             NA
#>   Number_of_technique_categories_used_in_palmitoyl_proteomics_studies
#> 1                                                                   0
#> 2                                                                  NA
#> 3                                                                  NA
#> 4                                                                  NA
#>   Technique_categories_used_in_palmitoyl_proteomics_studies
#> 1                                                        NA
#> 2                                                        NA
#> 3                                                        NA
#> 4                                                        NA
#>   Number_of_targeted_studies Targeted_studies__PMIDs PATs APTs Number_of_sites
#> 1                          0                      NA   NA   NA               0
#> 2                         NA                      NA   NA   NA              NA
#> 3                         NA                      NA   NA   NA              NA
#> 4                         NA                      NA   NA   NA              NA
#>   Sites_in_main_isoform Number_of_isoforms Max_number_of_cysteines
#> 1                    NA                  1                       3
#> 2                    NA                 NA                      NA
#> 3                    NA                 NA                      NA
#> 4                    NA                 NA                      NA
#>   Max_number_of_cysteines_in_TM_or_cytosolic_domain
#> 1                                                NA
#> 2                                                NA
#> 3                                                NA
#> 4                                                NA
#>   Predicted_to_be_S_palmitoylated
#> 1                           FALSE
#> 2                              NA
#> 3                              NA
#> 4                              NA
#>   Predicted_to_be_S_palmitoylated_in_cytosolic_domains
#> 1                                                FALSE
#> 2                                                   NA
#> 3                                                   NA
#> 4                                                   NA
#>   Protein_has_hits_in_SwissPalm
#> 1                         FALSE
#> 2                            NA
#> 3                            NA
#> 4                            NA
#>   Orthologs_of_this_protein_have_hits_in_SwissPalm    Found_in_SwissPalm
#> 1                                            FALSE                 found
#> 2                                               NA not found in database
#> 3                                               NA not found in database
#> 4                                               NA not found in database
# }