Retrieve palmitoylation data from the swissPalm database

Retrieve palmitoylation data from the swissPalm database

Usage

swissPalm(protein_id, dataset = "all", species = NULL)

Arguments

protein_id

Character vector with protein identifiers. Should be an ID type supported by SwissPalm (UniProt AC, UniProt secondary AC, UniProt ID, UniProt gene name, Ensembl protein, Ensembl gene, Refseq protein ID, IPI ID, UniGene ID, PomBase ID, MGI ID, RGD ID, TAIR protein ID, EuPathDb ID; from the SwissPalm website).

dataset

Which dataset to use in swissPalm. Will be set to "all" if not a value from swisspalm::datasets. Default = "all" (all datasets).

species

Which dataset to use in swissPalm. Will be set to "all" if not a value from swisspalm::species. Default = "all" (all datasets).

Value

Data frame with palmitoylation data for proteins in protein_id, and notes for proteins not found in SwissPalm.

Examples

swisspalmR::swissPalm(protein_id = c("P05067", "O00161", "P04899", "P98019"))
#>   Query identifier Found UniProt AC  UniProt ID   Organism
#> 1           O00161 found     O00161 SNP23_HUMAN H. sapiens
#> 2           P04899 found     P04899 GNAI2_HUMAN H. sapiens
#> 3           P05067 found     P05067    A4_HUMAN H. sapiens
#> 4           P98019 found     P98019  COX2_ANAPL    M. duck
#>                                  Protein name Found in palmitoyl-proteomes
#> 1 Synaptosomal-associated protein 23, SNAP-23                        20/22
#> 2        Guanine nucleotide-binding protein G                        17/22
#> 3         Amyloid-beta precursor protein, APP                         2/22
#> 4  Cytochrome c oxidase subunit 2, EC 7.1.1.9                          0/0
#>   Techniques Times validated
#> 1          4               4
#> 2          4               7
#> 3          2               4
#> 4          0               0
#>                                                    DHHC-PATs & APTs Sites
#> 1                                                                       6
#> 2 ZDHC3_MOUSE, ZDHC7_MOUSE, ZDHC2_MOUSE, ZDH21_MOUSE\n  LYPA1_HUMAN     3
#> 3                ZDHC7_MOUSE, ZDH21_MOUSE, ZDHC7_HUMAN, ZDH21_HUMAN     2
#> 4                                                                       0
#>   Predicted sites Cys Orthologs
#> 1             Yes   6       Yes
#> 2             Yes  11       Yes
#> 3             Yes  18       Yes
#> 4              No   3        No

# Use 'species' parameter to limit your results to those for a species of 
# interest
swisspalmR::swissPalm(
  protein_id = c("P05067", "O00161", "P04899", "P98019"),
  species = swisspalmR::species["Mallard duck"]
  )
#>   Query identifier                 Found UniProt AC UniProt ID Organism
#> 1           P98019                 found     P98019 COX2_ANAPL  M. duck
#> 2             <NA> not found in database       <NA>       <NA>     <NA>
#> 3             <NA> not found in database       <NA>       <NA>     <NA>
#> 4             <NA> not found in database       <NA>       <NA>     <NA>
#>                                 Protein name Found in palmitoyl-proteomes
#> 1 Cytochrome c oxidase subunit 2, EC 7.1.1.9                          0/0
#> 2                                       <NA>                         <NA>
#> 3                                       <NA>                         <NA>
#> 4                                       <NA>                         <NA>
#>   Techniques Times validated Sites Predicted sites Cys Orthologs
#> 1          0               0     0              No   3        No
#> 2         NA              NA    NA            <NA>  NA      <NA>
#> 3         NA              NA    NA            <NA>  NA      <NA>
#> 4         NA              NA    NA            <NA>  NA      <NA>
#>   Query.identifier
#> 1             <NA>
#> 2           O00161
#> 3           P04899
#> 4           P05067